# Add thread-local shared primitive values