--- a/crates/motoko/benches/bench_core.rs
+++ b/crates/motoko/benches/bench_core.rs
@@ -27,7 +27,7 @@ fn core_clone_larger(b: &mut Bencher) {
             x := x_ + 1;
             ?x_
         }}}}};
-        let i = Iter.range(3);
+        let i = Iter.range(100);
         var sum = 0;
         for (y in i) {
             sum := sum + 1;