--- a/crates/motoko/benches/bench_vector.rs
+++ b/crates/motoko/benches/bench_vector.rs
@@ -22,11 +22,12 @@ macro_rules! vec_clone_n {
 
             #[bench]
             fn shared_vec_clone(b: &mut Bencher) {
-                let vec: Shared<Vec<Shared<Value>>> = (0..$n)
-                    .into_iter()
-                    .map(|_| SystemTime::now().to_motoko().unwrap().share())
-                    .collect::<Vec<_>>()
-                    .share();
+                let vec: Shared<Vec<Shared<Value>>> = Shared::new(
+                    (0..$n)
+                        .into_iter()
+                        .map(|_| SystemTime::now().to_motoko().unwrap().share())
+                        .collect::<Vec<_>>(),
+                );
 
                 b.iter(|| vec.fast_clone())
             }