Before:
```
test eval_larger  ... bench:     737,916 ns/iter (+/- 21,795)
test eval_smaller ... bench:       2,225 ns/iter (+/- 21)
```

After:
```
test eval_larger  ... bench:     701,332 ns/iter (+/- 37,938)
test eval_smaller ... bench:       2,205 ns/iter (+/- 119)
```

(Essentially the same performance, but this change seems reasonable for reclaiming some extra memory).